Pet the CodeClimate
