+++ lib/agris/api/messages/query_changed_orders.rb
@@ -2,29 +2,12 @@
 module Agris
   module Api
     module Messages
-      class QueryChangedOrders < QueryBase
-        def initialize(time, detail)
-          @time = time
-          @detail = detail
-        end
-
+      class QueryChangedOrders < ChangedQueryBase
         def message_number
           80_900
         end
 
         protected
-
-        def input_hash
-          input_base_hash
-            .merge(
-              :@details => @detail,
-              locid: {
-                :@datetime => @time.strftime('%Y-%m-%dT%H:%M:%S'),
-                :@id => nil,
-                :@loccode => nil
-              }
-            )
-        end
 
         def xml_hash
           xml_base_hash
